Added a proper "more" icon and did some reorganizing.
