+++ app/src/main/java/com/codemate/koffeemate/views/UserQuickDialView.kt
@@ -16,24 +16,19 @@
 
 package com.codemate.koffeemate.views
 
-import android.animation.ObjectAnimator
 import android.content.Context
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.util.AttributeSet
-import android.view.animation.AccelerateDecelerateInterpolator
 import android.widget.FrameLayout
 import com.codemate.koffeemate.R
 import com.codemate.koffeemate.data.models.User
 import com.codemate.koffeemate.ui.userselector.UserItemAnimator
+import com.codemate.koffeemate.ui.userselector.UserQuickDialAdapter
 import com.codemate.koffeemate.ui.userselector.UserSelectListener
-import com.codemate.koffeemate.ui.userselector.UserSelectorAdapter
-import kotlinx.android.synthetic.main.recycler_item_user.view.*
 import kotlinx.android.synthetic.main.view_user_quick_dial.view.*
 import org.jetbrains.anko.backgroundResource
 import org.jetbrains.anko.dip
-import org.jetbrains.anko.imageResource
-import org.jetbrains.anko.onClick
 
 class UserQuickDialView : FrameLayout {
     constructor(ctx: Context) : super(ctx)
@@ -94,57 +89,5 @@ class UserQuickDialView : FrameLayout {
                     userSelectorAdapter.clear()
                 }
                 .start()
-    }
-}
-
-class UserQuickDialAdapter(
-        onUserSelectedListener: (User) -> Unit,
-        private val onMoreClickedListener: () -> Unit) : UserSelectorAdapter(onUserSelectedListener) {
-    val accelerateDecelerateInterpolator = AccelerateDecelerateInterpolator()
-
-    private val TYPE_USER = 1
-    private val TYPE_MORE = 2
-
-    override fun getItemCount() =
-            if (users.isEmpty()) 0
-            else users.size + 1
-
-    override fun getItemViewType(position: Int) =
-            if (position < users.size)
-                TYPE_USER
-            else
-                TYPE_MORE
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        when (getItemViewType(position)) {
-            TYPE_USER -> {
-                super.onBindViewHolder(holder, position)
-                holder.itemView.userName.text = users[position].profile.first_name
-
-            }
-            TYPE_MORE -> {
-                holder.itemView.profileImage.imageResource = R.mipmap.ic_launcher
-                holder.itemView.userName.text = holder.itemView.context.getString(R.string.more)
-                holder.itemView.onClick { onMoreClickedListener?.invoke() }
-            }
-        }
-
-        with(ObjectAnimator.ofFloat(1f, 0f)) {
-            interpolator = accelerateDecelerateInterpolator
-            duration = 750
-            repeatMode = ObjectAnimator.REVERSE
-            repeatCount = ObjectAnimator.INFINITE
-            startDelay = (position * 250).toLong()
-
-            addUpdateListener {
-                val value = animatedValue as Float
-
-                holder.itemView.apply {
-                    translationY = -value * 10
-                }
-            }
-
-            start()
-        }
     }
 }
